+++ tests/utils/test_shell_helper.py
@@ -105,6 +105,7 @@ class TestListCommands(ShellHelperTest):
         "delete",
         "edit",
         "export",
+        "git",
         "import",
         "init",
         "list",