+++ CHANGELOG.md
@@ -18,6 +18,7 @@ database (as suggested in #123), be sure to check out
 ### Added
 - the `-f` short-hand alias for the `--field` argument of the `open` command
 - the `opened_entries` attribute of the `OpenCommand` (which is accessible during the `PostOpenCommand` hook)
+- the new `git` command to simplify running git operations on the database (#124)
 
 ### Changed
 - an error will be logged when a file is not found during the `open` command